+++ tools/cooja/java/se/sics/cooja/contikimote/interfaces/ContikiBeeper.java
@@ -26,7 +26,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $Id: ContikiBeeper.java,v 1.7 2008/10/28 13:37:11 fros4943 Exp $
+ * $Id: ContikiBeeper.java,v 1.8 2008/12/04 14:03:42 joxe Exp $
  */
 
 package se.sics.cooja.contikimote.interfaces;
@@ -102,7 +102,7 @@ public class ContikiBeeper extends Beeper implements ContikiMoteInterface, Polle
   }
 
   private TimeEvent stopBeepEvent = new TimeEvent(0) {
-    public void execute(int t) {
+    public void execute(long t) {
       myEnergyConsumption = 0.0;
     }
   };
@@ -135,7 +135,7 @@ public class ContikiBeeper extends Beeper implements ContikiMoteInterface, Polle
           return;
         }
 
-        int currentTime = mote.getSimulation().getSimulationTime();
+        long currentTime = mote.getSimulation().getSimulationTime();
         statusLabel.setText("Last beep at time: " + currentTime);
 
         /* Beep on speakers */
